# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Atlas Treasury API
  version: 0.2.0
  description: Transfers, reserves, holds, reconciliation, and settlement operations
  contact:
    name: AtlasPayments Team
    url: https://github.com/shivemind/atlaspayments
servers:
  - url: http://localhost:3000
    description: Local development server. Override via Postman {{baseUrl}} environment variable.
...
paths:
...
  /api/v1/reserves/{id}:
    get:
      summary: Get reserve by ID
      description: Retrieves a single reserve by its ID.
      x-platform:
        kind: report
        entity: Reserve
        affectsLedger: false
      operationId: getReserveById
      tags:
        - Reserves
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Reserve details.
          content:
            application/json:
              schema:
                type: object
                required:
                  - reserve
                properties:
                  reserve:
                    type: object
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Reserve not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      summary: Update reserve
      description: Updates mutable fields on an existing reserve.
      x-platform:
        kind: transition
        entity: Reserve
        affectsLedger: false
        stateChanges:
          - metadata_updated
      operationId: updateReserve
      tags:
        - Reserves
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Reserve updated.
          content:
            application/json:
              schema:
                type: object
                required:
                  - reserve
                properties:
                  reserve:
                    type: object
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Reserve not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/holds:
    get:
      summary: List holds
      description: Returns a paginated list of holds for the authenticated merchant.
      x-platform:
        kind: report
        entity: Hold
        affectsLedger: false
      operationId: listHolds
      tags:
        - Reserves
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated list of holds.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create a hold
      description: Creates a hold to temporarily freeze funds for the authenticated merchant.
      x-platform:
        kind: transition
        entity: Hold
        affectsLedger: true
        stateChanges:
          - created
      operationId: createHold
      tags:
        - Reserves
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
                - currency
              properties:
                amount:
                  type: integer
                  minimum: 1
                currency:
                  type: string
                  minLength: 3
                  maxLength: 3
                reason:
                  type: string
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "201":
          description: Hold created.
          content:
            application/json:
              schema:
                type: object
                required:
                  - hold
                properties:
                  hold:
                    type: object
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Idempotency key conflict.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/holds/{id}:
    get:
      summary: Get hold by ID
      description: Retrieves a single hold by its ID.
      x-platform:
        kind: report
        entity: Hold
        affectsLedger: false
      operationId: getHoldById
      tags:
        - Reserves
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Hold details.
          content:
            application/json:
              schema:
                type: object
                required:
                  - hold
                properties:
                  hold:
                    type: object
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Hold not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      summary: Update hold
      description: Updates mutable fields on an existing hold.
      x-platform:
        kind: transition
        entity: Hold
        affectsLedger: false
        stateChanges:
          - metadata_updated
      operationId: updateHold
      tags:
        - Reserves
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Hold updated.
          content:
            application/json:
              schema:
                type: object
                required:
                  - hold
                properties:
                  hold:
                    type: object
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Hold not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/internal/recon/sources:
    get:
      summary: List reconciliation sources
      description: Returns a paginated list of reconciliation sources.
      x-platform:
        kind: resource
        entity: ReconSource
        affectsLedger: false
      operationId: listReconSources
      tags:
        - Reconciliation
      security:
        - PlatformAdminKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated list of reconciliation sources.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "401": &a1
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403": &a2
          description: Platform admin access required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create recon source
      description: Creates a new reconciliation source.
      x-platform:
        kind: resource
        entity: ReconSource
        affectsLedger: false
      operationId: createReconSource
      tags:
        - Reconciliation
      security:
        - PlatformAdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - type
              properties:
                name:
                  type: string
                type:
                  type: string
                config:
                  type: object
                metadata:
                  type: object
      responses:
        "201":
          description: Reconciliation source created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
        "400": &a3
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401": *a1
        "403": *a2
...
components:
  securitySchemes:
    MerchantKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for merchant-scoped operations. Pass as `Authorization: Bearer {{merchantApiKey}}`.
    PlatformAdminKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for platform admin operations. Pass as `Authorization: Bearer {{platformAdminKey}}`.
  schemas:
    ErrorResponse:
      description: Standard error response envelope returned for 4xx/5xx responses.
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code.
            message:
              type: string
              description: Human-readable error description.
    PaginationMeta:
      description: Pagination metadata included in list responses.
      type: object
      required:
        - page
        - pageSize
        - total
        - hasMore
      properties:
        page:
          type: integer
          description: Current page number (1-indexed).
        pageSize:
          type: integer
          description: Number of results per page.
        total:
          type: integer
          description: Total number of matching records.
        hasMore:
          type: boolean
          description: Whether additional pages exist.
...
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      description: |
        Client-generated unique key for safe request retries. Identical key + payload replays the original response.
      schema:
        type: string
        example: idem-key-a1b2c3d4
    PageParam:
      in: query
      name: page
      description: Page number (1-indexed).
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      in: query
      name: pageSize
      description: Number of results per page (max 100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    PathId:
      in: path
      name: id
      required: true
      description: Resource identifier.
      schema:
        type: string
...